third_party/WebKit/LayoutTests/webmidi/requestmidiaccess-expected.txt

Tests navigator.requestMIDIAccess.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.requestMIDIAccess().then(successCallback1, errorCallback1) did not throw exception.
PASS requestMIDIAccess() succeeded with access [object MIDIAccess].
PASS access.inputs().length is 1
PASS access.outputs().length is 1
PASS input attributes are correct
PASS output attributes are correct
PASS output.send([0xf1, 0x45, 0x7f]) threw exception SecurityError: Failed to execute 'send' on 'MIDIOutput': permission to send system exclusive messages is denied..
PASS requestMIDIAccess() was correctly blocked for System Exclusive access with error [object DOMError].
PASS successfullyParsed is true

TEST COMPLETE
